Relaced Telerik NumericUpDown controls with Syncfusion counterparts

+++ EAShow.Shared/Models/Mutation.cs
@@ -5,5 +5,5 @@ using ValueOf;
 
 namespace EAShow.Core.Core.Models
 {
-    public class Mutation : ValueOf<double, Mutation>{}
+    public class Mutation : ValueOf<decimal, Mutation>{}
 }
